feat(LottoMachine) LottoMachine은 시스템은 구매 금액에 상응하는 개수의 로또 번호를 만들어준다
LottoMachine은 시스템은 구매 금액에 상응하는 개수의 로또 번호를 만들어준다

+++ src/main/java/lotto/PrizeCalculator.java
@@ -12,6 +12,6 @@ public class PrizeCalculator {
 
         int count = result.values().stream().reduce(0, Integer::sum);
 
-        return (double)(total / (count * 1000L));
+        return total / (count * 1000.0d);
     }
 }
